Refactor `type_check_delineated_path` closer to old namespace behaviour
This refactors the `type_check_delineated_path` function to ensure that
namespace mutation matches the old behaviour.

Some re-organisation was required in order to avoid simultaneously
requiring unique access to both the namespace binding *and* the
module/enum_module borrowed from the namespace. Unfortunately, it seems
`instantiate_enum` requires mutable access to both the namespace *and*
the inner module separately. It needs to be reworked to avoid this
requirement (perhaps by accessing the inner module internally using a
path?). Until then, TODOs have been added to remove the `clone`.