Refactor `instantiate_enum` to not require aliasing namespace
Previously, `instantiate_enum` required two mutable references to
mutable namespaces. Seeing as both referred to parts of the same
namespace, this made it impossible to appease the borrow-checker, even
if both were used safely within the function.

This refactors the `instantiate_enum` function to take a `path` to its
module rather than a direct mutable reference, so that the reference can
be acquired only when needed.